packages/cli/src/templates/py/requirements.txt
suiloop>=0.1.0
python-dotenv
